requirements/local.txt
# Local development dependencies go here
-r base.txt

coverage==4.5.4
django-coverage-plugin==1.6.0

Sphinx==2.2.2
django-extensions==2.2.5
Werkzeug==0.16.0
django-test-plus==1.4.0
factory-boy==2.12.0

django-debug-toolbar==2.1

# improved REPL
ipdb==0.12.3

pytest-django==3.7.0
pytest-sugar==0.9.2
